CENG499/HW1/model/requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
